+++ zfit/pdf.py
@@ -23,6 +23,7 @@ __all__ = [
     "StudentT",
     "Gamma",
     "JohnsonSU",
+    "Landau",
     "Bernstein",
     "Chebyshev",
     "Legendre",
@@ -90,13 +91,7 @@ from .models.kde import (
     KDE1DimISJ,
 )
 from .models.morphing import SplineMorphingPDF
-from .models.physics import (
-    CrystalBall,
-    DoubleCB,
-    GaussExpTail,
-    GeneralizedCB,
-    GeneralizedGaussExpTail,
-)
+from .models.physics import CrystalBall, DoubleCB, GaussExpTail, GeneralizedCB, GeneralizedGaussExpTail, Landau
 from .models.polynomials import (
     Bernstein,
     Chebyshev,